feat(distance-based-range): unit 4d — 획득과 유지를 가른다. h 는 재서 정했다
경계에서 락이 진동하는 것을 막는다. **원칙: 「여기서 쏠 수 있나 · 멈춰도 되나」는 획득,
「이미 문 것을 놓나」는 유지.** 정지 판정에 유지 임계를 쓰면 적이 사거리 밖에서 멈춘다.

**`h = 0.1` 칸 — 추정이 아니라 측정이다.**
계측을 코드로 심었다: `RangePredicateMirrorTest` 가 「멈춘」 적이 프레임마다 실제로 얼마나
움직이는지를 잰다(멈춘 적도 밀어냄·분리로 흔들리고, 그 폭이 곧 판정이 뒤집힐 수 있는 폭이다).
실측 **0.047 · 0.051칸**(2회). 한 표본에 딱 맞추지 않았다 — 지터는 밀집도에 따라 커진다.
0.1 은 관측치의 약 2배이면서 **옛 슬랙 0.5 의 1/5** 다. 사거리 1 의 유지 임계가 1.5 가 아니라
1.1 이라 `target-persistence` D2(「지나쳐 갔는데 붙들고 있는」 상태 제거)를 되살리지 않는다.

- `TargetPersistence.KeepsLock` 이 히스테리시스를 소유하고 내부에서 `InReach(range + h)` 를
  부른다 — 두 축이 **같은 술어**를 쓰되 임계만 다르다. 그래서 `InReach` 의 `tileRange` 가
  `int`→`float` 이 됐다.
- 유지 5곳(적 focus 락 · frontmost 락 · 방어유닛 focus 락 · committed 재판정 ·
  `EnemyAiStateSystem` 락 미러) / 나머지는 획득 그대로.
- 부수 정리: 유지 지점들이 `InReach`(몸 거리)와 `KeepsLock`(셀 체비셰프)을 **둘 다** 걸고
  있었다. 두 자가 겹쳐 실질 임계가 「더 엄한 쪽」이었고, 이제 하나다.
- `h` 는 코드 상수 — `sceneKnobs` 미등재(계약 9). 등재하면 `configHash` 가 움직여 골든 red 가
  「조건 드리프트」로 읽힌다.

관측: `force_wave` 이벤트 482→**487**(락이 조금 더 유지돼 공격이 몇 번 더 나갔다),
`long_boss` 2줄, 나머지 6건 무변화. 킬/유출 전건 무변화 — **판 전체를 안 바꾸는 것이
이 unit 의 성공 조건**이고 그대로 나왔다.

⚠ **완료 기준 1 은 부분 달성이다.** 「관측한 진동이 사라진다」를 직접 확인하지 못했다 —
카나리아가 보는 것은 「멈췄는데 안 쏨」(교착)이지 「쏘다 말다」(진동)가 아니다. 잰 것은 진동의
**원인**(위치 지터)이고 h 가 그 폭을 덮는 것은 산술로 참이지만, 육안 확인은 unit 5(사거리 원
표시)가 도구를 만든 뒤로 미룬다. spec 에 그대로 적었다.

검증: EditMode 2667건 / 실패 2건(선행 — `boomerang`·`bomb_man` 문안) ·
PlayMode 교착 카나리아 통과.

+++ Assets/_Project/Scripts/Battle/Combat/AttackReach.cs
@@ -55,7 +55,9 @@ namespace Wassup.Battle.Combat
         //
         // ⚠ **`tileSize` 로 나눠 타일 단위로 넘긴다.** 술어가 월드 단위를 모르는 이유는
         // 「사거리 3」이 저작에서 타일 수이기 때문이다 — 월드로 환산하는 지점이 하나여야 한다.
-        public static bool InReach(float3 atkPos, float3 tgtPos, int tileRange,
+        // ⚠ `tileRange` 가 **실수**인 이유: 유지 판정이 히스테리시스 폭 `h` 를 더해 부른다
+        // (`TargetPersistence.KeepsLock`). 저작은 정수지만 술어는 그걸 알 필요가 없다.
+        public static bool InReach(float3 atkPos, float3 tgtPos, float tileRange,
                                    float tileSize, float targetBodyRadiusTiles = 0f)
         {
             float inv = tileSize > 1e-6f ? 1f / tileSize : 1f;
